+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.4] - 2025-10-31
+
+### Added
+- Added a `Switch-ByteNibble` private function and removed the Conversion dependency
+
+### Fixed
+- Had some *stupid*, *STUPID* mistakes in the Pester tests. 100% passing now.
+
 ## [2.8.3] - 2025-10-30
 
 ### Added